backend/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Authentication service provider built with FastAPI"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.135.3",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "ruff>=0.15.11",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
